Fix various PHP warnings, refs #13178
Fixed a number of PHP warnings, some of which were introduced with PHP
7.2.

+++ apps/qubit/modules/actor/actions/editAction.class.php
@@ -226,32 +226,42 @@ class ActorEditAction extends DefaultEditAction
       case 'subjectAccessPoints':
       case 'placeAccessPoints':
         $value = $filtered = array();
-        foreach ($this->form->getValue($field->getName()) as $item)
-        {
-          $params = $this->context->routing->parse(Qubit::pathInfo($item));
-          $resource = $params['_sf_route']->resource;
-          $value[$resource->id] = $filtered[$resource->id] = $resource;
-        }
 
-        foreach ($this[$field->getName()] as $item)
+        if (is_array($formItems = $this->form->getValue($field->getName())))
         {
-          if (isset($value[$item->term->id]))
+          foreach ($formItems as $item)
           {
-            unset($filtered[$item->term->id]);
-          }
-          else
-          {
-            $item->indexObjectOnDelete = false;
-            $item->delete();
+            $params = $this->context->routing->parse(Qubit::pathInfo($item));
+            $resource = $params['_sf_route']->resource;
+            $value[$resource->id] = $filtered[$resource->id] = $resource;
           }
         }
 
-        foreach ($filtered as $item)
+        if (is_array($this[$field->getName()]))
         {
-          $relation = new QubitObjectTermRelation;
-          $relation->term = $item;
+          foreach ($this[$field->getName()] as $item)
+          {
+            if (isset($value[$item->term->id]))
+            {
+              unset($filtered[$item->term->id]);
+            }
+            else
+            {
+              $item->indexObjectOnDelete = false;
+              $item->delete();
+            }
+          }
+        }
 
-          $this->resource->objectTermRelationsRelatedByobjectId[] = $relation;
+        if (is_array($filtered))
+        {
+          foreach ($filtered as $item)
+          {
+            $relation = new QubitObjectTermRelation;
+            $relation->term = $item;
+
+            $this->resource->objectTermRelationsRelatedByobjectId[] = $relation;
+          }
         }
 
         break;
